Conflicts:
	.github/workflows/ci.yml
	.github/workflows/ci_linux/build_pyscf.sh
	.github/workflows/ci_linux/python_deps.sh
	.github/workflows/ci_macos/python_deps.sh
	pyscf/cc/test/test_eom_uccsd.py
	pyscf/lib/CMakeLists.txt
	pyscf/lib/dft/CMakeLists.txt
	pyscf/lib/dft/libxc_itrf.c
	pyscf/lib/gto/ft_ao.h
	pyscf/lib/linalg_helper.py
	pyscf/lib/pbc/CMakeLists.txt
	pyscf/lib/pbc/optimizer.c
	pyscf/lib/test/test_numint_uniform_grid.py
	pyscf/mp/test/test_gmp2.py
	pyscf/mp/test/test_mp2.py
	pyscf/pbc/cc/test/test_eom_krccsd.py
	pyscf/pbc/ci/test/test_ci.py
	pyscf/pbc/dft/multigrid/multigrid.py
	pyscf/pbc/dft/test/test_multigrid.py
	pyscf/pbc/grad/test/test_krhf.py
	pyscf/pbc/tools/pbc.py
	pyscf/scf/atom_hf.py

+++ pyscf/pbc/lib/kpts_helper.py
@@ -148,11 +148,11 @@ def check_kpt_antiperm_symmetry(array, idx1, idx2, tolerance=1e-8):
         True
     '''
     # Checking to make sure bounds of idx1 and idx2 are O.K.
-    assert(idx1 >= 0 and idx2 >= 0 and 'indices to swap must be non-negative!')
+    assert (idx1 >= 0 and idx2 >= 0 and 'indices to swap must be non-negative!')
 
     array_shape_len = len(array.shape)
     nparticles = (array_shape_len + 1) / 4
-    assert(idx1 < (2 * nparticles - 1) and idx2 < (2 * nparticles - 1) and
+    assert (idx1 < (2 * nparticles - 1) and idx2 < (2 * nparticles - 1) and
            'This function does not support the swapping of the last k-point index '
            '(This k-point is implicitly not indexed due to conservation of momentum '
            'between k-points.).')
@@ -209,7 +209,7 @@ def get_kconserv3(cell, kpts, kijkab):
                 kconserv[i,j,mask] = c
 
     new_shape = [shape[i] for i, x in enumerate(kijkab)
-                 if not isinstance(x, (int,np.int))]
+                 if not isinstance(x, (int, np.integer))]
     kconserv = kconserv.reshape(new_shape)
     return kconserv
